+++ src/util/json_helper.cpp
@@ -123,7 +123,7 @@ ghoul::Dictionary jsonToDictionary(const nlohmann::json& json) {
                 // We can't represent arrays with different types, so we have to use a
                 // Dictionary for that instead
                 ghoul::Dictionary subDict;
-                for (int i = 0; i < j.size(); i++) {
+                for (size_t i = 0; i < j.size(); i++) {
                     const nlohmann::json& value = j[i];
                     // We add 1 to the key to make Lua happy :-/
                     addToDict(subDict, std::format("{}", i + 1), value);
